// packages/tuya-panel-dragon-sdk/src/interface.ts
export interface Color {
  hue: number;
  saturation: number;
  value: number;
}

export interface WhiteColor {
  brightness: number; // 白光亮度
  temperature: number; // 白光色温
}

export enum SchemaType {
  RAW = 'raw',
  Boolean = 'bool',
  String = 'string',
  Enum = 'enum',
  Number = 'value',
  BitMap = 'bitmap',
  Unknow = 'unknow',
}

export type SchemaDataType = string | number | boolean | Color | any;

export interface DpSchema {
  code: string;
  desc?: string;
  dptype?: string;
  iconname?: string;
  id: number;
  max?: number;
  min?: number;
  mode: 'rw' | 'wr';
  name?: string;
  scale?: number;
  step?: number;
  range?: string[];
  maxlen?: number; // string type 下的最大长度
  type: SchemaType;
  validate?: (value: SchemaDataType) => boolean;
  format?: (value: SchemaDataType) => SchemaDataType;
}

export interface DpSchemas {
  [key: string]: DpSchema;
}

export interface CloudSchema {}
// ...
export interface IFormater {
  uuid: string;
  schema: DpSchema | CloudSchema;

  /**
   * 比较两个值是否一致
   */
  equal: (source: SchemaDataType, target: SchemaDataType) => boolean;

  /**
   * 将标准协议数据转为项目数据
   */
  parse: (value: SchemaDataType) => SchemaDataType;

  /**
   * 将数据转为标准协议数据
   */
  format: (value: SchemaDataType) => SchemaDataType;
}